Enhanced Date Handling, Improved Terraform Config, and Added Failsafe Notebook
Resolved issues related to date fetch and update from the Wikimedia API. The fixes address the incompatibilities that were previously causing errors during data extraction.
Enhanced the robustness of date functions to handle irregular input more effectively. This improvement ensures that the data extraction process can handle a wider range of date formats and values without breaking.
Updated the Terraform configuration to reflect all changes across the infrastructure and architecture. This update ensures that the Terraform configuration accurately represents the current state of the infrastructure.
Updated the .gitignore file to exclude sensitive files. This update will help prevent accidental commit of sensitive information in future code updates.
Added a local backfill notebook as a failsafe measure and for local run capabilities. This notebook can be used to manually extract and upload data to BigQuery in case of any disruptions in the main pipeline. It can also be used for local testing and development.